api/src/app/requirements.txt
fastapi==0.78.0
uvicorn==0.17.6
uvloop==0.16.0
orjson==3.6.7
PyJWT==2.4.0
backoff==2.0.1
pydantic==1.9.1
python-dotenv==0.20.0
loguru==0.6.0
psycopg2-binary==2.9.3